requirements.txt
pandas
requests
pytesseract
pillow
opencv-python
openpyxl
flet